git checkout -b fix-slash-commands-tm4j (#7397)
Automatic Merge
